uclibc: upstream changes to unused files
This commit summarizes changes by upstream to files, architectures or
subsystems not used by L4Re as well as the upstream buildsystem

* unused architectures

csky: Add EM_CSKY for ldso
or1k: Fix compiling with PIC and latest binutils use PLT for __syscall_error
PowerPC: Fix termios definitions
sparc: remove asm constraint
sparc64: Use the jmpbuf-offsets.h header instead of duplicating it
sparc64: Make structure match kernel uapi
sparc64: Fix value for O_mask_was_saved based on gdb observation
xtensa: add custom bits/poll.h
xtensa: don't use l32r opcode explicitly

* unused features / files / subsystems

Defined INLINE_SYSCALL_NOERR_NCS in mips/bits/syscalls.h
several patches for uclibc ng
fix issues in ethers.c
Fix NULL pointer dereference in open_memstream()
preadv/pwritev: fix offset argument type
Fix compilation issue when libuargp is compiled without __UCLIBC_HAS_PROGRAM_INVOCATION_NAME__
Make __syscall_error return long, as expected by syscall() callers
implement fexecve from glibc
fix PTRAVE_EVENT_SECCOMP typo in ptrace.h
poll: avoid calling select with empty sets which hangs the process

* upstream buildsystem changes (upstream buildsystem not used by L4re)

MAINTAINERS: add Mao Han for CSKY
docs/defines.txt: clarify LFS-related macros
Fix header installation for recent 64-only arch
Makefile.in: build utils without stack smashing protector
aarch64: always use MMU

+++ uclibc/lib/contrib/uclibc/libc/sysdeps/linux/alpha/__syscall_error.c
@@ -8,7 +8,7 @@
 
 /* This routine is jumped to by all the syscall handlers, to stash
    an error number into errno.  */
-int attribute_hidden __syscall_error (void)
+long attribute_hidden __syscall_error (void)
 {
 	register int err_no __asm__("$0");
 	__set_errno (err_no);
